requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
requests
redis>=5.0